Fix warnings raised by extremely picky MSVC 1911
git-svn-id: https://svn.osgeo.org/gdal/trunk@41120 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/mrf/libLERC/Huffman.cpp
@@ -63,7 +63,7 @@ bool Huffman::ComputeCodes(const vector<int>& histo)
 
   m_codeTable.resize(size);
   std::fill( m_codeTable.begin(), m_codeTable.end(),
-             std::pair<short, unsigned int>(0, 0) );
+             std::pair<short, unsigned int>(static_cast<short>(0), 0) );
 
   if (!pq.top().TreeToLUT(0, 0, m_codeTable))    // fill the LUT
     return false;
